activated JS modules
