+++ index.html
@@ -54,6 +54,6 @@
         <p class="footer-links">Privacy policy | Cookie information | Terms of use</p>
         <p>Fait avec soin par mes petites mains ♥ Petites Mains 2024</p>
      </footer>
-     <script src="main.js"></script>
+     <script src="main.js" type="module"></script>
 </body>
 </html>
